+++ src/fem/block/element/composite/localization/wedge/mod.rs
@@ -49,7 +49,128 @@ where
     {
         self.calculate_deformation_gradient_rates_composite_localization_element(nodal_coordinates, nodal_velocities)
     }
-    composite_surface_element_boilerplate_inner!{}
+    fn calculate_inverse_normalized_projection_matrix() -> NormalizedProjectionMatrix<Q>
+    {
+        let diag: Scalar = 3.0/64.0;
+        let off: Scalar = -1.0/64.0;
+        NormalizedProjectionMatrix::new([
+            [diag,  off,  off],
+            [ off, diag,  off],
+            [ off,  off, diag]
+        ])
+    }
+    fn calculate_projected_gradient_vectors(_reference_nodal_coordinates: &ReferenceNodalCoordinates<O>) -> ProjectedGradientVectors<G, N>
+    {
+        panic!()
+    }
+    fn calculate_reference_jacobians_subelements(reference_nodal_coordinates: &ReferenceNodalCoordinates<O>) -> Scalars<P>
+    {
+        Self::calculate_bases(reference_nodal_coordinates).iter()
+        .map(|reference_basis_vectors|
+            reference_basis_vectors[0].cross(&reference_basis_vectors[1]).norm()
+        ).collect()
+    }
+    fn calculate_shape_function_integrals() -> ShapeFunctionIntegrals<P, Q>
+    {
+        ShapeFunctionIntegrals::new([
+            [32.0,  8.0,  8.0],
+            [ 8.0, 32.0,  8.0],
+            [ 8.0,  8.0, 32.0],
+            [16.0, 16.0, 16.0]
+        ])
+    }
+    fn calculate_shape_function_integrals_products() -> ShapeFunctionIntegralsProducts<P, Q>
+    {
+        ShapeFunctionIntegralsProducts::new([[
+            [22.0,  5.0,  5.0],
+            [ 5.0,  2.0,  1.0],
+            [ 5.0,  1.0,  2.0]
+        ], [
+            [ 2.0,  5.0,  1.0],
+            [ 5.0, 22.0,  5.0],
+            [ 1.0,  5.0,  2.0]
+        ], [
+            [ 2.0,  1.0,  5.0],
+            [ 1.0,  2.0,  5.0],
+            [ 5.0,  5.0, 22.0]
+        ], [
+            [ 6.0,  5.0,  5.0],
+            [ 5.0,  6.0,  5.0],
+            [ 5.0,  5.0,  6.0]
+        ]])
+    }
+    fn calculate_shape_functions_at_integration_points() -> ShapeFunctionsAtIntegrationPoints<G, Q>
+    {
+        let diag: Scalar = 0.666_666_666_666_666_6;
+        let off: Scalar = 0.166_666_666_666_666_7;
+        ShapeFunctionsAtIntegrationPoints::new([
+            [diag,  off,  off],
+            [ off, diag,  off],
+            [ off,  off, diag]
+        ])
+    }
+    fn calculate_standard_gradient_operators() -> StandardGradientOperators<M, O, P>
+    {
+        StandardGradientOperators::new([[
+            [ 2.0,  0.0],
+            [ 0.0,  0.0],
+            [ 0.0,  0.0],
+            [ 0.0,  2.0],
+            [ 0.0,  0.0],
+            [-2.0, -2.0]
+        ], [
+            [ 0.0,  0.0],
+            [ 0.0,  2.0],
+            [ 0.0,  0.0],
+            [ 2.0,  0.0],
+            [-2.0, -2.0],
+            [ 0.0,  0.0]
+        ], [
+            [ 0.0,  0.0],
+            [ 0.0,  0.0],
+            [-2.0, -2.0],
+            [ 0.0,  0.0],
+            [ 0.0,  2.0],
+            [ 2.0,  0.0]
+        ], [
+            [ 0.0,  0.0],
+            [ 0.0,  0.0],
+            [ 0.0,  0.0],
+            [ 2.0,  2.0],
+            [-2.0,  0.0],
+            [ 0.0, -2.0]
+        ]])
+    }
+    fn calculate_standard_gradient_operators_transposed() -> StandardGradientOperatorsTransposed<M, O, P>
+    {
+        let standard_gradient_operators = Self::calculate_standard_gradient_operators();
+        let mut standard_gradient_operators_transposed = StandardGradientOperatorsTransposed::zero();
+        standard_gradient_operators_transposed.iter_mut().enumerate()
+        .for_each(|(n, standard_gradient_operators_transposed_n)|
+            standard_gradient_operators_transposed_n.iter_mut()
+            .zip(standard_gradient_operators.iter())
+            .for_each(|(standard_gradient_operators_transposed_n_e, standard_gradient_operators_e)|
+                standard_gradient_operators_transposed_n_e.iter_mut()
+                .zip(standard_gradient_operators_e[n].iter())
+                .for_each(|(standard_gradient_operators_transposed_n_e_i, standard_gradient_operators_e_n_i)|
+                    *standard_gradient_operators_transposed_n_e_i = *standard_gradient_operators_e_n_i
+                )
+            )
+        );
+        standard_gradient_operators_transposed
+    }
+    fn get_constitutive_models(&self) -> &[C; G]
+    {
+        &self.constitutive_models
+    }
+    fn get_integration_weights(&self) -> &Scalars<G>
+    {
+        &self.integration_weights
+    }
+    fn get_projected_gradient_vectors(&self) -> &ProjectedGradientVectors<G, N>
+    {
+        &self.projected_gradient_vectors
+    }
 }
 
 impl<'a, C> CompositeLocalizationElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
@@ -424,4 +545,71 @@ where
     }
 }
 
-composite_surface_or_localization_element_boilerplate!(Wedge);
+impl<'a, C> CompositeSurfaceElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
+where
+    C: Constitutive<'a>
+{
+    fn get_scaled_reference_normals(&self) -> &ScaledReferenceNormals<G, P>
+    {
+        &self.scaled_reference_normals
+    }
+}
+
+impl<'a, C> ElasticCompositeElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
+where
+    C: Elastic<'a>
+{}
+
+impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for Wedge<C>
+where
+    C: Hyperelastic<'a>
+{
+    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
+    {
+        self.calculate_helmholtz_free_energy_composite_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> HyperelasticCompositeElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
+where
+    C: Hyperelastic<'a>
+{}
+
+impl<'a, C> ViscoelasticCompositeElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
+where
+    C: Viscoelastic<'a>
+{}
+
+impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for Wedge<C>
+where
+    C: ElasticHyperviscous<'a>
+{
+    fn calculate_viscous_dissipation(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
+    {
+        self.calculate_viscous_dissipation_composite_element(nodal_coordinates, nodal_velocities)
+    }
+    fn calculate_dissipation_potential(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
+    {
+        self.calculate_dissipation_potential_composite_element(nodal_coordinates, nodal_velocities)
+    }
+}
+
+impl<'a, C> ElasticHyperviscousCompositeElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
+where
+    C: ElasticHyperviscous<'a>
+{}
+
+impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for Wedge<C>
+where
+    C: Hyperviscoelastic<'a>
+{
+    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
+    {
+        self.calculate_helmholtz_free_energy_composite_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> HyperviscoelasticCompositeElement<'a, C, G, M, N, O, P, Q> for Wedge<C>
+where
+    C: Hyperviscoelastic<'a>
+{}
