+++ src/fem/block/element/composite/tetrahedron/mod.rs
@@ -343,4 +343,89 @@ where
     }
 }
 
-super::composite_element_boilerplate!(Tetrahedron);
+impl<'a, C> ElasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+where
+    C: Elastic<'a>
+{
+    fn calculate_nodal_forces(&self, nodal_coordinates: &NodalCoordinates<N>) -> NodalForces<N>
+    {
+        self.calculate_nodal_forces_composite_element(nodal_coordinates)
+    }
+    fn calculate_nodal_stiffnesses(&self, nodal_coordinates: &NodalCoordinates<N>) -> NodalStiffnesses<N>
+    {
+        self.calculate_nodal_stiffnesses_composite_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> ElasticCompositeElement<'a, C, G, M, N, O, P, Q> for Tetrahedron<C>
+where
+    C: Elastic<'a>
+{}
+
+impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+where
+    C: Hyperelastic<'a>
+{
+    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
+    {
+        self.calculate_helmholtz_free_energy_composite_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> HyperelasticCompositeElement<'a, C, G, M, N, O, P, Q> for Tetrahedron<C>
+where
+    C: Hyperelastic<'a>
+{}
+
+impl<'a, C> ViscoelasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+where
+    C: Viscoelastic<'a>
+{
+    fn calculate_nodal_forces(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> NodalForces<N>
+    {
+        self.calculate_nodal_forces_composite_element(nodal_coordinates, nodal_velocities)
+    }
+    fn calculate_nodal_stiffnesses(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> NodalStiffnesses<N>
+    {
+        self.calculate_nodal_stiffnesses_composite_element(nodal_coordinates, nodal_velocities)
+    }
+}
+
+impl<'a, C> ViscoelasticCompositeElement<'a, C, G, M, N, O, P, Q> for Tetrahedron<C>
+where
+    C: Viscoelastic<'a>
+{}
+
+impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for Tetrahedron<C>
+where
+    C: ElasticHyperviscous<'a>
+{
+    fn calculate_viscous_dissipation(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
+    {
+        self.calculate_viscous_dissipation_composite_element(nodal_coordinates, nodal_velocities)
+    }
+    fn calculate_dissipation_potential(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
+    {
+        self.calculate_dissipation_potential_composite_element(nodal_coordinates, nodal_velocities)
+    }
+}
+
+impl<'a, C> ElasticHyperviscousCompositeElement<'a, C, G, M, N, O, P, Q> for Tetrahedron<C>
+where
+    C: ElasticHyperviscous<'a>
+{}
+
+impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+where
+    C: Hyperviscoelastic<'a>
+{
+    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
+    {
+        self.calculate_helmholtz_free_energy_composite_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> HyperviscoelasticCompositeElement<'a, C, G, M, N, O, P, Q> for Tetrahedron<C>
+where
+    C: Hyperviscoelastic<'a>
+{}
