+++ src/fem/block/element/linear/localization/wedge/mod.rs
@@ -47,11 +47,34 @@ where
     {
         self.calculate_deformation_gradient_rate_linear_localization_element(nodal_coordinates, nodal_velocities)
     }
+    fn calculate_gradient_vectors(_: &ReferenceNodalCoordinates<O>) -> GradientVectors<N>
+    {
+        panic!()
+    }
     fn calculate_reference_jacobian(reference_nodal_coordinates_midplane: &ReferenceNodalCoordinates<O>) -> Scalar
     {
         Self::calculate_reference_jacobian_linear_surface_element(reference_nodal_coordinates_midplane)
     }
-    linear_surface_element_boilerplate_inner!{}
+    fn calculate_standard_gradient_operator() -> StandardGradientOperator<M, O>
+    {
+        StandardGradientOperator::new([
+            [-1.0, -1.0],
+            [ 1.0,  0.0],
+            [ 0.0,  1.0]
+        ])
+    }
+    fn get_constitutive_model(&self) -> &C
+    {
+        &self.constitutive_model
+    }
+    fn get_gradient_vectors(&self) -> &GradientVectors<N>
+    {
+        &self.gradient_vectors
+    }
+    fn get_integration_weight(&self) -> &Scalar
+    {
+        &self.integration_weight
+    }
 }
 
 impl<'a, C> LinearLocalizationElement<'a, C, G, M, N, O> for Wedge<C>
@@ -273,4 +296,71 @@ where
     }
 }
 
-super::linear_surface_or_localization_element_boilerplate!(Wedge);
+impl<'a, C> LinearSurfaceElement<'a, C, G, M, N, O> for Wedge<C>
+where
+    C: Constitutive<'a>
+{
+    fn get_reference_normal(&self) -> &ReferenceNormal
+    {
+        &self.reference_normal
+    }
+}
+
+impl<'a, C> ElasticLinearElement<'a, C, G, M, N, O> for Wedge<C>
+where
+    C: Elastic<'a>
+{}
+
+impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for Wedge<C>
+where
+    C: Hyperelastic<'a>
+{
+    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
+    {
+        self.calculate_helmholtz_free_energy_linear_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> HyperelasticLinearElement<'a, C, G, M, N, O> for Wedge<C>
+where
+    C: Hyperelastic<'a>
+{}
+
+impl<'a, C> ViscoelasticLinearElement<'a, C, G, M, N, O> for Wedge<C>
+where
+    C: Viscoelastic<'a>
+{}
+
+impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for Wedge<C>
+where
+    C: ElasticHyperviscous<'a>
+{
+    fn calculate_viscous_dissipation(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
+    {
+        self.calculate_viscous_dissipation_linear_element(nodal_coordinates, nodal_velocities)
+    }
+    fn calculate_dissipation_potential(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
+    {
+        self.calculate_dissipation_potential_linear_element(nodal_coordinates, nodal_velocities)
+    }
+}
+
+impl<'a, C> ElasticHyperviscousLinearElement<'a, C, G, M, N, O> for Wedge<C>
+where
+    C: ElasticHyperviscous<'a>
+{}
+
+impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for Wedge<C>
+where
+    C: Hyperviscoelastic<'a>
+{
+    fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
+    {
+        self.calculate_helmholtz_free_energy_linear_element(nodal_coordinates)
+    }
+}
+
+impl<'a, C> HyperviscoelasticLinearElement<'a, C, G, M, N, O> for Wedge<C>
+where
+    C: Hyperviscoelastic<'a>
+{}
