+++ src/fem/block/element/linear/surface/mod.rs
@@ -223,212 +223,3 @@ where
     }
     fn get_reference_normal(&self) -> &ReferenceNormal;
 }
-
-macro_rules! linear_surface_element_boilerplate
-{
-    ($element: ident) =>
-    {
-        impl<'a, C> ElasticFiniteElement<'a, C, G, N> for $element<C>
-        where
-            C: Elastic<'a>
-        {
-            fn calculate_nodal_forces(&self, nodal_coordinates: &NodalCoordinates<N>) -> NodalForces<N>
-            {
-                self.calculate_nodal_forces_linear_element(nodal_coordinates)
-            }
-            fn calculate_nodal_stiffnesses(&self, nodal_coordinates: &NodalCoordinates<N>) -> NodalStiffnesses<N>
-            {
-                let first_piola_kirchoff_tangent_stiffness = self.get_constitutive_model()
-                .calculate_first_piola_kirchoff_tangent_stiffness(
-                    &self.calculate_deformation_gradient(nodal_coordinates)
-                );
-                let gradient_vectors = self.get_gradient_vectors();
-                let identity = TensorRank2::<3, 1, 1>::identity();
-                let normal_gradients = Self::calculate_normal_gradients(nodal_coordinates);
-                let reference_normal = self.get_reference_normal();
-                gradient_vectors.iter()
-                .map(|gradient_vector_a|
-                    gradient_vectors.iter()
-                    .zip(normal_gradients.iter())
-                    .map(|(gradient_vector_b, normal_gradient_b)|
-                        first_piola_kirchoff_tangent_stiffness.iter()
-                        .map(|first_piola_kirchoff_tangent_stiffness_m|
-                            identity.iter()
-                            .zip(normal_gradient_b.iter())
-                            .map(|(identity_n, normal_gradient_b_n)|
-                                first_piola_kirchoff_tangent_stiffness_m.iter()
-                                .zip(gradient_vector_a.iter())
-                                .map(|(first_piola_kirchoff_tangent_stiffness_mj, gradient_vector_a_j)|
-                                    first_piola_kirchoff_tangent_stiffness_mj.iter()
-                                    .zip(identity_n.iter()
-                                    .zip(normal_gradient_b_n.iter()))
-                                    .map(|(first_piola_kirchoff_tangent_stiffness_mjk, (identity_nk, normal_gradient_b_n_k))|
-                                        first_piola_kirchoff_tangent_stiffness_mjk.iter()
-                                        .zip(gradient_vector_b.iter()
-                                        .zip(reference_normal.iter()))
-                                        .map(|(first_piola_kirchoff_tangent_stiffness_mjkl, (gradient_vector_b_l, reference_normal_l))|
-                                            first_piola_kirchoff_tangent_stiffness_mjkl * gradient_vector_a_j * (
-                                                identity_nk * gradient_vector_b_l + normal_gradient_b_n_k * reference_normal_l
-                                            ) * self.get_integration_weight()
-                                        ).sum::<Scalar>()
-                                    ).sum::<Scalar>()
-                                ).sum::<Scalar>()
-                            ).collect()
-                        ).collect()
-                    ).collect()
-                ).collect()
-            }
-        }
-        impl<'a, C> ViscoelasticFiniteElement<'a, C, G, N> for $element<C>
-        where
-            C: Viscoelastic<'a>
-        {
-            fn calculate_nodal_forces(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> NodalForces<N>
-            {
-                self.calculate_nodal_forces_linear_element(nodal_coordinates, nodal_velocities)
-            }
-            fn calculate_nodal_stiffnesses(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> NodalStiffnesses<N>
-            {
-                let first_piola_kirchoff_rate_tangent_stiffness = self.get_constitutive_model()
-                .calculate_first_piola_kirchoff_rate_tangent_stiffness(
-                    &self.calculate_deformation_gradient(nodal_coordinates),
-                    &self.calculate_deformation_gradient_rate(nodal_coordinates, nodal_velocities)
-                );
-                let gradient_vectors = self.get_gradient_vectors();
-                let identity = TensorRank2::<3, 1, 1>::identity();
-                let normal_gradients = Self::calculate_normal_gradients(nodal_coordinates);
-                let reference_normal = self.get_reference_normal();
-                gradient_vectors.iter()
-                .map(|gradient_vector_a|
-                    gradient_vectors.iter()
-                    .zip(normal_gradients.iter())
-                    .map(|(gradient_vector_b, normal_gradient_b)|
-                        first_piola_kirchoff_rate_tangent_stiffness.iter()
-                        .map(|first_piola_kirchoff_rate_tangent_stiffness_m|
-                            identity.iter()
-                            .zip(normal_gradient_b.iter())
-                            .map(|(identity_n, normal_gradient_b_n)|
-                                first_piola_kirchoff_rate_tangent_stiffness_m.iter()
-                                .zip(gradient_vector_a.iter())
-                                .map(|(first_piola_kirchoff_rate_tangent_stiffness_mj, gradient_vector_a_j)|
-                                    first_piola_kirchoff_rate_tangent_stiffness_mj.iter()
-                                    .zip(identity_n.iter()
-                                    .zip(normal_gradient_b_n.iter()))
-                                    .map(|(first_piola_kirchoff_rate_tangent_stiffness_mjk, (identity_nk, normal_gradient_b_n_k))|
-                                        first_piola_kirchoff_rate_tangent_stiffness_mjk.iter()
-                                        .zip(gradient_vector_b.iter()
-                                        .zip(reference_normal.iter()))
-                                        .map(|(first_piola_kirchoff_rate_tangent_stiffness_mjkl, (gradient_vector_b_l, reference_normal_l))|
-                                            first_piola_kirchoff_rate_tangent_stiffness_mjkl * gradient_vector_a_j * (
-                                                identity_nk * gradient_vector_b_l + normal_gradient_b_n_k * reference_normal_l
-                                            ) * self.get_integration_weight()
-                                        ).sum::<Scalar>()
-                                    ).sum::<Scalar>()
-                                ).sum::<Scalar>()
-                            ).collect()
-                        ).collect()
-                    ).collect()
-                ).collect()
-            }
-        }
-        linear_surface_or_localization_element_boilerplate!($element);
-    }
-}
-pub(crate) use linear_surface_element_boilerplate;
-
-macro_rules! linear_surface_or_localization_element_boilerplate
-{
-    ($element: ident) =>
-    {
-        impl<'a, C> LinearSurfaceElement<'a, C, G, M, N, O> for $element<C>
-        where
-            C: Constitutive<'a>
-        {
-            fn get_reference_normal(&self) -> &ReferenceNormal
-            {
-                &self.reference_normal
-            }
-        }
-        impl<'a, C> ElasticLinearElement<'a, C, G, M, N, O> for $element<C>
-        where
-            C: Elastic<'a>
-        {}
-        impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for $element<C>
-        where
-            C: Hyperelastic<'a>
-        {
-            fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
-            {
-                self.calculate_helmholtz_free_energy_linear_element(nodal_coordinates)
-            }
-        }
-        impl<'a, C> HyperelasticLinearElement<'a, C, G, M, N, O> for $element<C>
-        where
-            C: Hyperelastic<'a>
-        {}
-        impl<'a, C> ViscoelasticLinearElement<'a, C, G, M, N, O> for $element<C>
-        where
-            C: Viscoelastic<'a>
-        {}
-        impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for $element<C>
-        where
-            C: ElasticHyperviscous<'a>
-        {
-            fn calculate_viscous_dissipation(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
-            {
-                self.calculate_viscous_dissipation_linear_element(nodal_coordinates, nodal_velocities)
-            }
-            fn calculate_dissipation_potential(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> Scalar
-            {
-                self.calculate_dissipation_potential_linear_element(nodal_coordinates, nodal_velocities)
-            }
-        }
-        impl<'a, C> ElasticHyperviscousLinearElement<'a, C, G, M, N, O> for $element<C>
-        where
-            C: ElasticHyperviscous<'a>
-        {}
-        
-        impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for $element<C>
-        where
-            C: Hyperviscoelastic<'a>
-        {
-            fn calculate_helmholtz_free_energy(&self, nodal_coordinates: &NodalCoordinates<N>) -> Scalar
-            {
-                self.calculate_helmholtz_free_energy_linear_element(nodal_coordinates)
-            }
-        }
-        impl<'a, C> HyperviscoelasticLinearElement<'a, C, G, M, N, O> for $element<C>
-        where
-            C: Hyperviscoelastic<'a>
-        {}
-    }
-}
-pub(crate) use linear_surface_or_localization_element_boilerplate;
-
-macro_rules! linear_surface_element_boilerplate_inner
-{
-    () =>
-    {
-        fn calculate_standard_gradient_operator() -> StandardGradientOperator<M, O>
-        {
-            StandardGradientOperator::new([
-                [-1.0, -1.0],
-                [ 1.0,  0.0],
-                [ 0.0,  1.0]
-            ])
-        }
-        fn get_constitutive_model(&self) -> &C
-        {
-            &self.constitutive_model
-        }
-        fn get_gradient_vectors(&self) -> &GradientVectors<N>
-        {
-            &self.gradient_vectors
-        }
-        fn get_integration_weight(&self) -> &Scalar
-        {
-            &self.integration_weight
-        }
-    }
-}
-pub(crate) use linear_surface_element_boilerplate_inner;
